requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
dj-database-url==2.0.0
Django==4.0.5
django-allauth==0.54.0
django-ckeditor==6.5.1
django-environ==0.10.0
django-etc==1.4.0
django-hitcount==1.3.5
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-modeltranslation==0.18.9
django-social-share==2.3.0
environs==9.5.0
gunicorn==20.1.0
idna==3.4
marshmallow==3.19.0
oauthlib==3.2.2
packaging==23.1
Pillow==9.1.1
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.30.0
requests-oauthlib==1.3.1
sqlparse==0.4.2
typing-extensions==4.5.0
urllib3==2.0.2
whitenoise==6.5.0
